Fixed `ls` to to work on ".", "..".  Fixes #25
`getDirectoryPermissions` does not work on directories that end with a
trailing slash.  Unfortunately, `Filesystem.Path.decodeString` adds a
trailing slash to the end of the "." and ".." paths.  Also, if you
provide a path with a trailing slash it will fail, too.  This fix
uses a new `deslash` helper function to sanitize paths for
`getDirectoryPermissions`