Improve exception propagation for subprocess runners.  Fixes #113
Subprocess runners all have some sort of functionality to feed a stream to
their input handle.  However, if the standard input stream throws an
exception they fail to close the input handle which in some cases causes
the process to hang indefinitely, such as `cat`.

A really simple example of this problem is the following program:

    {-# LANGUAGE OverloadedStrings #-}

    import Turtle

    main = do
        let stream1 = inproc "fdsdfd" [] empty
           stream2 = inproc "cat" [] stream1
        stdout stream2

This program would previously hang because `inproc` would throw an
exception and then wouldn't close the input handle to the `cat`
subprocess, which would hang indefinitely waiting for input.

The solution is to modify the `feedIn` function within each subprocess
runner to use `finally` to ensure that the handle is always properly
closed in the face of an exception.