Change `rmtree` on Unix-symlinks. Fixes #143
Before `rmtree` would call `Filesystem.isDirectory` on the symlink, which
resolves the symlink _before_ testing whether it is a directory.
If the symlink refers to the directory, then it would delete all the contents
of that directory and afterwards fail to remove the symlink itself.

Now it behaves like `rm -r` on unix systems. When called on a symlink it just
deletes that symlink irregardles of what it points to.