Pass the `--reorder-goals` flag to `cabal` in CI
Cabal has difficulty finding a solution when compiling both the tests and the
benchmarks at the same time on older versions of `ghc`.  This is due to the
following issue:

* `doctest` depends on `ghc-7.6.3`/`ghc-7.8.2` which in turn depends on
  `binary-0.5.1.1`
* `criterion-0.6.2.1` is the newest version compatible with `binary-0.5.1.1`
* `cabal` gives up before `criterion-0.6.2.1`, due to `--max-backjumps`

Even if you enable unlimited backtracking (i.e. `--max-backjumps=-1`) `cabal`
still takes forever to find a solution, whereas `--reorder-goals` finds a
solution nearly instantly, so I prefer that approach.