Use GHC version macros instead of MIN_VERSION_base
The ghci shipped with ghc-7.10 and earlier do not pick up cabal
generated version macros like `MIN_VERSION_base`. For older GHC, the
`__GLASGOW_HASKELL__` macro is available and since we're only testing
the version of the base library `__GLASGOW_HASKELL__` is a good
substitute because the base library version is intimately tied to
the GHC version.

Tested with ghc 7.8.4, 7.10.3 and 8.0.1.