Fix build failures on older versions of `ghc`
The `Turtle.Bytes.strict` function doesn't compile against older version of
`ghc` due to using `fmap` without specifying a `Functor` constraint.  This is
because `MonadIO` does not have a `Functor` super class for older versions of
GHC, which leads to a build failure.  The fix in this case is to either use
`liftM` or `do` notation and this fix uses `do` notation.