Fix subprocess commands to not force all input.  Fixes #178
Many subprocess runners take a `Shell` as input, but they would `wait` for
the input stream to run to completion before terminating.  That meant that if
you ran something like:

```
{-# LANGUAGE OverloadedStrings #-}

import Turtle

main :: IO ()
main = proc "echo" ["foo"] stdin
```

... then `echo` would print `foo`, but would also wait for `stdin` to run to
completion.  Furthermore, if you provided one line of standard input then
`stdin` would fail with a broken pipe exception due to `echo` closing its
input handle.

The solution is to replace every `wait` with a newly added internal `halt`
function.  Now if the downstream process is done it will `poll` the upstream
process to see if it's done.  If the upstream process is not done it will just
`cancel` the process and if the upstream process failed with an exception we
rethrow the exception to preserve it.