Fix test failure
Adding the `Turtle.Line` import in the `Turtle.Format` module caused the
doc-tests to fail.  The reason why is that you can only import other
modules from the same package if they are also being doc-tested, but
`Turtle.Line` was not being doc-tested, which led to an import failure
when testing `Turtle.Format`.

However, if you just add `Turtle.Line` to the list of modules to test
that in turn leads to a test failure since some of the examples in the
`Turtle.Line` module do not pass unless you import other `turtle`
modules.  I decided to disable those doctests rather than import further
modules.