Change behavior of `textToLines ""`.  Fixes #229
This changes `textToLines` to behave like `Data.Text.splitOn "\n"`
instead of `Data.Text.unlines`.  The reason why is that the former
behaves more consistently in its treatment of the empty string.

For example:

```
>>> Data.Text.lines "\na\na"
["","a","a"]
>>> Data.Text.lines "\na"
["","a"]
>>> Data.Text.lines ""
[]
>>> Data.Text.lines "\na\na"
["","a","a"]
>>> Data.Text.lines "\na"
["","a"]
>>> Data.Text.lines ""
[]
```

... or:

```haskell
>>> Data.Text.lines "aa"
["aa"]
>>> Data.Text.lines "a"
["a"]
>>> Data.Text.lines ""
[]
>>> Data.Text.splitOn "\n" "aa"
["aa"]
>>> Data.Text.splitOn "\n" "a"
["a"]
>>> Data.Text.splitOn "\n" ""
[""]
```

One good example of how `splitOn` is the desired behavior is in the bug
report that prompted this fix.  Returning an empty list caused `sed` to
misbehave on empty lines, whereas returning a 1-element line of an
empty string restored the correct behavior.