Add `MonadCatch` instance for `Shell`
Fixes #259

This is a breaking change that requires changing the internal type that
`Shell` uses from `FoldM` to `FoldIO`.  In practice, most code should
work without any issues and the code that is affected can be fixed by
just replacing `Shell` with `_Shell`.

This also slightly changes the semantics of `FoldM`-based folds by
lazily deferring the @begin@ step of the `FoldM` until the first
@step@