Add Turtle.Format.makePrinter function
I was trying to implement some abstraction on top of Turtle to redirect
the default output (stdout and/or stderr) to a logger function withing
some monad where the logger is available. While I was able to implement
most of the Turtle function in one way or another it turned out it’s not
that trivial to implement custom `printf` and `eprintf` since `(>>-)` of
`Turtle.Format.Format` is not exposed.

This change adds `makePrinter` function that just adds extra argument
for `printf` that is a printer function ending with `m ()` as `Format`
result. The implementation is trivial (`flip (>>-)`), `printf` and
`eprintf` are reimplemented using this function.