Restore `collapse` utility
It is now implemented in terms of the `normalise` function

This behavior is *not* the same as the old `collapse` utility.
Specifically, the way they handle `..` is different, but I decided
that the difference in behavior is an improvement and therefore
opted to not try to exactly preserve the old behavior.